components: model `Button[F]#clicks` as a view over `Discord[F]#events`
modeling `clicks` as a view over `Discord[F]#events` is more natural!

this means that `Interaction[F]` no longer has to worry about "waiting
for future component interactions and dispatching them to the
`Component[F]`s"; that responsibility is offloaded to the `Cmd[F]`, and
components no longer get "dispatched" to (this is imperative brain at
play). this also allows for natural cancelation propagation and
discharging the `Concurrent` bound from `Interaction.withEvent`
(intuitively, lifting an event into an interaction algebra doesn't
necessarily imply something has to occur concurrently).

`InteractionResponse[F]` has also become effectful due to now storing
the components (themselves effectful) that were sent to Discord
alongside their generated IDs. we accept the response when instantiating
the clicks stream in order to match the component ID back to the button
and only react to relevant events. this also enforces that the component
_was actually sent_ in the first place (i.e. can't make a button and
immediately try to listen for clicks on it without sending it to Discord
first).